Removed casts

+++ Core2D.UI.Uwp/Renderer/Win2dRenderer.cs
@@ -43,10 +43,10 @@ namespace Test.Uwp
         private Color ToColor(ArgbColor color)
         {
             return Color.FromArgb(
-                (byte)color.A,
-                (byte)color.R,
-                (byte)color.G,
-                (byte)color.B);
+                color.A,
+                color.R,
+                color.G,
+                color.B);
         }
 
         private static CanvasStrokeStyle CreateStrokeStyle(BaseStyle style)
