+++ Core2D.UI.Wpf/Converters/ArgbColorToBrushMultiConverter.cs
@@ -26,10 +26,10 @@ namespace Core2D.UI.Wpf.Converters
             {
                 var brush = new SolidColorBrush(
                     Color.FromArgb(
-                        (byte)((int)values[0]),
-                        (byte)((int)values[1]),
-                        (byte)((int)values[2]),
-                        (byte)((int)values[3])));
+                        (byte)values[0],
+                        (byte)values[1],
+                        (byte)values[2],
+                        (byte)values[3]));
                 brush.Freeze();
                 return brush;
             }
